+++ lib/extensions/object.dart
@@ -9,23 +9,34 @@ extension ObjectExtension on Object? {
 
   bool equals(dynamic compare) => this != null && this == compare;
 
-  T? get<T>(String key, [T? defaultValue]) {
+  T find<T extends Object?>({Object? key, T? defaultValue}) {
+    final T? arguments = findOrNull(key: key, defaultValue: defaultValue);
+    if (arguments != null) {
+      return arguments;
+    } else {
+      throw UnimplementedError("$T didn't find from this object");
+    }
+  }
+
+  T findByKey<T extends Object?>(String key, T defaultValue) {
+    return find(key: key, defaultValue: defaultValue);
+  }
+
+  T? findOrNull<T extends Object?>({Object? key, T? defaultValue}) {
     var root = this;
-    if (root is Map<String, dynamic>) {
-      var data = root[key];
-      if (data is T) {
-        return data;
+    if (root is Map) {
+      var arguments = root[key];
+      if (arguments is T) {
+        return arguments;
       } else {
         return defaultValue;
       }
     } else {
-      if (root is T) {
+      if (key == null && root is T) {
         return root;
       } else {
         return defaultValue;
       }
     }
   }
-
-  T findByKey<T>(String key, T defaultValue) => get(key) ?? defaultValue;
 }